Add license to include OR public domain, split out
This commit splits the License portion of the README out into a separate file,
and adds an OR clause specifying that the Public Domain license is also
acceptable. This allows a more explicit license statement that keeps with the
spirit of the original statement of license for those who find this facilitates
their use of the work.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>